refactor(plugin-assets): less duplication
