--- packages/plugin-assets/src/errors.js
@@ -1,14 +1,0 @@
-const createErrorHandler = (debug, spinner) => {
-  return (error, message) => {
-    if (message) {
-      spinner.fail(message);
-    } else {
-      spinner.stop();
-    }
-    debug('0%', error);
-  };
-};
-
-module.exports = {
-  createErrorHandler,
-};
